+++ x-pack/plugins/cases/server/services/configure/index.ts
@@ -5,8 +5,6 @@
  * 2.0.
  */
 
-import Boom from '@hapi/boom';
-
 import type {
   Logger,
   SavedObject,
@@ -15,10 +13,7 @@ import type {
 } from '@kbn/core/server';
 
 import { ACTION_SAVED_OBJECT_TYPE } from '@kbn/actions-plugin/server';
-import type {
-  ConfigurationAttributes,
-  CustomFieldsConfiguration,
-} from '../../../common/types/domain';
+import type { ConfigurationAttributes } from '../../../common/types/domain';
 import { CONNECTOR_ID_REFERENCE_NAME } from '../../common/constants';
 import { decodeOrThrow } from '../../../common/api';
 import { CASE_CONFIGURE_SAVED_OBJECT } from '../../../common/constants';
@@ -44,32 +39,6 @@ import {
   ConfigurationPartialAttributesRt,
   ConfigurationTransformedAttributesRt,
 } from '../../common/types/configure';
-
-/**
- * Throws an error if the requests has custom fields with duplicated keys.
- */
-function throwIfDuplicatedCustomFieldKeysInRequest({
-  customFieldsInRequest = [],
-}: {
-  customFieldsInRequest?: CustomFieldsConfiguration;
-}) {
-  const uniqueKeys = new Set();
-  const duplicatedKeys = new Set();
-
-  customFieldsInRequest.forEach((item) => {
-    if (uniqueKeys.has(item.key)) {
-      duplicatedKeys.add(item.key);
-    } else {
-      uniqueKeys.add(item.key);
-    }
-  });
-
-  if (duplicatedKeys.size) {
-    throw Boom.badRequest(
-      `Invalid duplicated custom field keys in request: ${Array.from(duplicatedKeys.values())}`
-    );
-  }
-}
 
 export class CaseConfigureService {
   constructor(private readonly log: Logger) {}
@@ -151,10 +120,6 @@ export class CaseConfigureService {
 
       const decodedAttributes = decodeOrThrow(ConfigurationTransformedAttributesRt)(attributes);
 
-      throwIfDuplicatedCustomFieldKeysInRequest({
-        customFieldsInRequest: decodedAttributes.customFields,
-      });
-
       const esConfigInfo = transformAttributesToESModel(decodedAttributes);
 
       const createdConfig =
@@ -184,10 +149,6 @@ export class CaseConfigureService {
       this.log.debug(`Attempting to UPDATE case configuration ${configurationId}`);
 
       const decodedAttributes = decodeOrThrow(ConfigurationPartialAttributesRt)(updatedAttributes);
-
-      throwIfDuplicatedCustomFieldKeysInRequest({
-        customFieldsInRequest: updatedAttributes.customFields,
-      });
 
       const esUpdateInfo = transformAttributesToESModel(decodedAttributes);
